ATM consultancy refactored
