+++ app/src/main/java/com/matheus/androidudemycourse/feature/atm_consultancy/fragments/CustomerFragment.kt
@@ -1,0 +1,22 @@
+package com.matheus.androidudemycourse.feature.atm_consultancy.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.matheus.androidudemycourse.databinding.FragmentCustomerBinding
+
+
+class CustomerFragment : Fragment() {
+
+    private lateinit var binding: FragmentCustomerBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentCustomerBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+}
