+++ app/src/main/java/com/matheus/androidudemycourse/feature/atm_consultancy/fragments/ServicesFragment.kt
@@ -1,11 +1,10 @@
-package com.matheus.androidudemycourse.feature.atm_consultancy.fragments.services
+package com.matheus.androidudemycourse.feature.atm_consultancy.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.matheus.androidudemycourse.R
 import com.matheus.androidudemycourse.databinding.FragmentServicesBinding
 
 class ServicesFragment : Fragment() {
