+++ app/src/main/res/layout/activity_atm_consultancy.xml
@@ -9,7 +9,7 @@
     tools:openDrawer="start">
 
     <include
-        android:id="@+id/aapBarMainAtm"
+        android:id="@+id/app_bar_main_atm"
         layout="@layout/app_bar_atm"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
