+++ app/src/main/res/layout/fragment_atm_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".feature.atm_consultancy.fragments.main.AtmMainFragment">
+    tools:context=".feature.atm_consultancy.fragments.AtmMainFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/logoAtmImageView"
