+++ app/src/main/res/layout/fragment_contact_atm.xml
@@ -3,5 +3,5 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.atm_consultancy.fragments.contact.ContactAtmFragment">
+    tools:context=".feature.atm_consultancy.fragments.ContactAtmFragment">
 </androidx.constraintlayout.widget.ConstraintLayout>
