+++ app/src/main/res/layout/fragment_customer.xml
@@ -4,32 +4,33 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.atm_consultancy.fragments.clients.ClientsFragment">
+    tools:context=".feature.atm_consultancy.fragments.CustomerFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
-        android:id="@+id/clientsTitleTextView"
+        android:id="@+id/customerTitleTextView"
         android:layout_width="wrap_content"
+        app:layout_constraintBottom_toTopOf="@id/customerDescriptionTextView"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="25dp"
-        android:text="@string/atm_clients"
+        android:text="@string/atm_customer"
         android:textSize="25sp"
         android:textStyle="bold"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <androidx.appcompat.widget.AppCompatTextView
-        android:id="@+id/clientsDescriptionTextView"
+        android:id="@+id/customerDescriptionTextView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:text="@string/clients_text"
+        android:text="@string/customer_text"
         android:textSize="20sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/clientsTitleTextView" />
+        app:layout_constraintTop_toBottomOf="@id/customerTitleTextView" />
 
     <androidx.appcompat.widget.AppCompatImageView
         android:layout_width="0dp"
@@ -42,5 +43,5 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/clientsDescriptionTextView" />
+        app:layout_constraintTop_toBottomOf="@id/customerDescriptionTextView" />
 </androidx.constraintlayout.widget.ConstraintLayout>
