+++ app/src/main/res/menu/activity_atm_drawer.xml
@@ -15,7 +15,7 @@
         <item
             android:id="@+id/nav_clients_atm"
             android:icon="@drawable/ic_clients"
-            android:title="@string/atm_clients" />
+            android:title="@string/atm_customer" />
     </group>
 
     <item android:title="@string/submenu_title">
